+++ application/dns/filter/domains.go
@@ -2,21 +2,55 @@ package gatesentryDnsFilter
 
 import (
 	"bufio"
+	"encoding/json"
 	"log"
 	"net/http"
 	"strings"
 	"sync"
+	"time"
 
 	gatesentry2storage "bitbucket.org/abdullah_irfan/gatesentryf/storage"
+	gatesentryTypes "bitbucket.org/abdullah_irfan/gatesentryf/types"
 )
 
-func InitializeFilters(blockedDomains *map[string]bool, blockedLists *[]string, internalRecords *map[string]string, exceptionDomains *map[string]bool, mutex *sync.Mutex, settings *gatesentry2storage.MapStore) {
+func InitializeFilters(blockedDomains *map[string]bool, blockedLists *[]string, internalRecords *map[string]string, exceptionDomains *map[string]bool, mutex *sync.Mutex, settings *gatesentry2storage.MapStore, dnsinfo *gatesentryTypes.DnsServerInfo) {
+	*blockedDomains = make(map[string]bool)
+	*blockedLists = []string{}
+	*internalRecords = make(map[string]string)
+	*exceptionDomains = make(map[string]bool)
+	dnsinfo.NumberDomainsBlocked = 0
+	custom_entries := settings.Get("dns_custom_entries")
+	log.Println("[DNS.SERVER] Custom entries found")
+	// unmarshall json array string to array
+	custom_entries_array := []string{}
+	//convert string to byte array
+	err := json.Unmarshal([]byte(custom_entries), &custom_entries_array)
+	if err != nil {
+		log.Println("[DNS.SERVER] Error unmarshalling custom entries:", err)
+	} else {
+		// check if blocklists already contains custom entries
+		entriesAdded := 0
+		for _, custom_entry := range custom_entries_array {
+			found := false
+			for _, blocklist := range *blockedLists {
+				if blocklist == custom_entry {
+					found = true
+					break
+				}
+			}
+			if !found {
+				*blockedLists = append(*blockedLists, custom_entry)
+				entriesAdded++
+			}
+		}
+		log.Println("[DNS.SERVER] Custom entries added to blocklists count:", entriesAdded)
+	}
 	InitializeInternalRecords(internalRecords, mutex, settings)
-	InitializeBlockedDomains(blockedDomains, blockedLists, mutex)
+	InitializeBlockedDomains(blockedDomains, blockedLists, mutex, dnsinfo)
 	InitializeExceptionDomains(exceptionDomains, mutex)
 }
 
-func InitializeBlockedDomains(blockedDomains *map[string]bool, blocklists *[]string, mutex *sync.Mutex) {
+func InitializeBlockedDomains(blockedDomains *map[string]bool, blocklists *[]string, mutex *sync.Mutex, dnsinfo *gatesentryTypes.DnsServerInfo) {
 	var wg sync.WaitGroup
 	log.Println("[DNS] Downloading blocklists...")
 
@@ -29,9 +63,10 @@ func InitializeBlockedDomains(blockedDomains *map[string]bool, blocklists *[]str
 				log.Println("[DNS] [Error] Failed to fetch blocklist:", err)
 				return
 			}
-			addDomainsToBlockedMap(blockedDomains, domains, mutex)
+			addDomainsToBlockedMap(blockedDomains, domains, mutex, dnsinfo)
 		}(blocklistURL)
 	}
+	dnsinfo.LastUpdated = int(time.Now().Unix())
 
 	wg.Wait()
 	log.Println("[DNS] Blocklists downloaded and processed.")
@@ -75,12 +110,13 @@ func fetchDomainsFromBlocklist(url string) ([]string, error) {
 	return domains, nil
 }
 
-func addDomainsToBlockedMap(blockedDomains *map[string]bool, newDomains []string, mutex *sync.Mutex) {
+func addDomainsToBlockedMap(blockedDomains *map[string]bool, newDomains []string, mutex *sync.Mutex, dnsinfo *gatesentryTypes.DnsServerInfo) {
 	mutex.Lock()
 	defer mutex.Unlock()
 
 	for _, domain := range newDomains {
 		(*blockedDomains)[domain] = true
+		dnsinfo.NumberDomainsBlocked++
 	}
 
 	log.Println("[DNS] Added", len(newDomains), "domains to blocked map")